refactor(gui): move the MQTT import page and the device templates out of flow.ts
The import page reads a broker rather than the PDU, is used once when wiring something up, and has nothing
to do with drawing a diagram. It needs two things from the flow code — the config migration and the save —
and nothing needs it back.

Device templates get their own module because two pages instantiate them: the import page and the Nodes
page's "Import device" panel. Instantiating one writes real config — a Modbus connection, pre-wired nodes
and links — so two slightly different ways of doing it would produce two slightly different devices.

flow.ts 3121 -> 2749. Every module but the import page now depends on nothing from it.
