+++ rPDU2MQTT.Web/web/build.mjs
@@ -26,6 +26,7 @@ const MODULES = [
   'energy.ts',
   'history-control.ts',
   'charts.ts',
+  'node-templates.ts',
   'palette.ts',
   'overrides.ts',
   'sections/paths.ts',
@@ -34,6 +35,7 @@ const MODULES = [
   'sections/livedata.ts',
   'sections/flow.ts',
   'sections/energy-board.ts',
+  'sections/mqtt-import.ts',
   'sections/nodedata.ts',
   'sections/trends.ts',
   'sections/export.ts',
